Fixed issues with ASP.NET metadata providers
